scripts/scrapers/requirements.txt
# DEX v2.0 — Python Scraper Requirements
# Install with: pip install -r requirements.txt

python-jobspy
requests
beautifulsoup4
playwright
google-generativeai
supabase
python-dotenv
lxml
